# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Orders API",
    "version": "1.0.0",
    "description": "API simples para gerenciamento de pedidos (desafio Jitterbit)."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },

    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "productId": { "type": "integer", "example": 2434 },
          "quantity": { "type": "integer", "example": 1 },
          "price": { "type": "number", "example": 1000 }
        }
      },

      "Order": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string", "example": "v1008991vdb-01" },
          "value": { "type": "number", "example": 10000 },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "example": "2023-07-19T12:24:11.529Z"
          },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Item" }
          }
        }
      }
    }
  },

  "paths": {
    "/login": {
      "post": {
        "summary": "Gera um token JWT simples para testes",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"],
                "example": {
                  "username": "admin",
                  "password": "123456"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token gerado com sucesso",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "example": {
                  "message": "Credenciais inválidas"
                }
              }
            }
          }
        }
      }
    },

    "/order": {
      "post": {
        "summary": "Cria um novo pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["numeroPedido", "valorTotal", "dataCriacao", "items"],
                "properties": {
                  "numeroPedido": { "type": "string" },
                  "valorTotal": { "type": "number" },
                  "dataCriacao": { "type": "string", "format": "date-time" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "idItem": { "type": "string" },
                        "quantidadeItem": { "type": "integer" },
                        "valorItem": { "type": "number" }
                      }
                    }
                  }
                },
                "example": {
                  "numeroPedido": "v1008991vdb-01",
                  "valorTotal": 10000,
                  "dataCriacao": "2023-07-19T12:24:11.529Z",
                  "items": [
                    {
                      "idItem": "2434",
                      "quantidadeItem": 1,
                      "valorItem": 1000
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pedido criado" },
          "400": { "description": "Dados inválidos" },
          "409": { "description": "Pedido já existe" }
        }
      }
    },

    "/order/{orderId}": {
      "get": {
        "summary": "Busca um pedido pelo orderId",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "404": { "description": "Pedido não encontrado" }
        }
      },

      "put": {
        "summary": "Atualiza um pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": { "type": "number" },
                  "creationDate": { "type": "string", "format": "date-time" },
                  "items": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Item" }
                  }
                },
                "example": {
                  "value": 20000,
                  "creationDate": "2023-07-19T12:24:11.529Z",
                  "items": [
                    {
                      "productId": 2434,
                      "quantity": 2,
                      "price": 1000
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pedido atualizado" },
          "400": { "description": "Body inválido" },
          "404": { "description": "Pedido não encontrado" }
        }
      },

      "delete": {
        "summary": "Remove um pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido removido",
            "content": {
              "application/json": {
                "example": { "message": "Pedido removido com sucesso" }
              }
            }
          },
          "404": { "description": "Pedido não encontrado" }
        }
      }
    },

    "/order/list": {
      "get": {
        "summary": "Lista todos os pedidos",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      }
    }
  }
}